components/backend/ems/pyproject.toml
[project]
name = "ems"
version = "0.1.0"
description = "Event management system"
authors = [
    { name = "Evgeny Mangasaryan", email = "mangasaryan.ep@gmail.com" },
    { name = "Dmitry Mikhailin", email = "aboba@aboba.org"},
]
dependencies = [
    "aiofiles>=23.2.1",
    "alembic>=1.11.1",
    "asyncio>=3.4.3",
    "asyncpg>=0.28.0",
    "attrs>=23.1.0",
    "fastapi>=0.104.1",
    "gunicorn>=21.2.0",
    "Pillow>=10.1.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.2",
    "python-dateutil>=2.8.2",
    "python-jose>=3.3.0",
    "python-json-logger>=2.0.7",
    "python_multipart>=0.0.6",
    "PyYAML>=6.0.1",
    "SQLAlchemy>=2.0.19",
    "uvicorn>=0.23.1",
]
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.3.5",
    "mypy>=1.9.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ems", "src/ems_libs"]

[tool.ruff]
line-length = 79
target-version = "py312"
exclude = [
    "*__init__.py",
    "tests/"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
ignore_missing_imports = true

[tool.pylsp-mypy]
enabled = true
live_mode = true
strict = false
